src/requirements.txt
flask
flask_sqlalchemy
python-dotenv
flask-restx
flask-marshmallow
marshmallow-sqlalchemy
flask_jwt_extended